==> notebooks/2. Merging.ipynb <==
### Merge with IMDB ratings dataset

# Import Criterion dataset
import pandas as pd
data = pd.read_csv('data\Criterion.csv')

# Convert year to string
data['Year'] = data['Year'].astype(str)

# Combine title and year for matching
data['titleYear'] = data['Title'] + ' ' + data['Year']

len(data)

data.head()

# Import IMDB dataset
ratings = pd.read_csv('title.ratings.tsv', sep = '\t')
basics = pd.read_csv('title.basics.tsv', sep = '\t', low_memory = False)
basics = basics.drop(['titleType', 'isAdult', 'endYear', 'runtimeMinutes', 'genres'], axis = 1)

# Merge to obtain ratings + titles
imdb_merged = ratings.merge(basics, left_on = 'tconst', right_on = 'tconst')

# Combine title and year for matching
imdb_merged['titleYear'] = imdb_merged['primaryTitle'] + ' ' + imdb_merged['startYear']

len(imdb_merged)

# With duplicate titleYear, keep the one with the most numVotes
imdb_merged = imdb_merged.sort_values(['titleYear', 'numVotes'], ascending = False)
imdb_merged['isDuplicated'] = imdb_merged.duplicated('titleYear', keep = 'first')
imdb_merged = imdb_merged[imdb_merged['isDuplicated'] == False]

len(imdb_merged)

imdb_merged.head()

# Sanity check
imdb_merged[imdb_merged['titleYear'] == 'Weekend 2011']

### TD-IDF Cosine Similarity String Matching
https://github.com/Bergvca/string_grouper

# Import string_grouper
from string_grouper import match_strings, \
match_most_similar, group_similar_strings, \
compute_pairwise_similarities, StringGrouper

To do: Find a way to match on both primaryTitle and originalTitle

# Match on titleYear
matches = match_most_similar(imdb_merged['titleYear'], data['titleYear'])

matches

# Drop NAs (no match)
nas = matches[matches['most_similar_index'].isna()]
nas = nas['most_similar_titleYear']
matches = matches.dropna(axis = 0)

# Check NAs
nas

matches.head()

# Remove NAs from data
import warnings
warnings.filterwarnings("ignore", 'This pattern has match groups')
data = data[~data['titleYear'].str.contains('|'.join(nas))]

# Rename Criterion titles based on IMDB titles
pd.options.mode.chained_assignment = None
data['newTitleYear'] = ''
for i in range(0, data.index[-1] + 1):
    try:
        index = int(matches.loc[i, 'most_similar_index'])
    except:
        pass
    data.loc[i, 'newTitleYear'] = imdb_merged.loc[index, 'titleYear']

len(data)

data = data.dropna(axis = 0)

len(data)

data

criterion_merged = data.merge(imdb_merged, left_on = 'newTitleYear', right_on = 'titleYear')

len(criterion_merged)

criterion_merged.sort_values('averageRating', ascending = False)

# Save to csv
criterion_merged.to_csv('data\Merged.csv', index = False)
